refactor: designate moved vo and firstcollection directory
