+++ BE/src/main/java/com/codesquad/issue04/domain/issue/vo/firstcollection/Comments.java
@@ -1,4 +1,4 @@
-package com.codesquad.issue04.domain.issue.vo;
+package com.codesquad.issue04.domain.issue.vo.firstcollection;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -8,6 +8,7 @@ import javax.persistence.Embeddable;
 import javax.persistence.FetchType;
 import javax.persistence.OneToMany;
 
+import com.codesquad.issue04.domain.issue.vo.Comment;
 import lombok.Getter;
 
 @Getter
